Adjust to latest compliant hardware specs (#882)
* 60MHz ram

* RGB565 framebuffer

+++ sources/Adapters/adv/Core/Inc/ltdc.h
@@ -24,8 +24,8 @@ extern "C" {
 /* USER CODE BEGIN Includes */
 #define DISPLAY_WIDTH 720
 #define DISPLAY_HEIGHT 720
-extern uint32_t framebuffer[];
-extern uint32_t placeholder_framebuffer_layer2[];
+extern uint16_t framebuffer[];
+extern uint16_t placeholder_framebuffer_layer2[];
 /* USER CODE END Includes */
 
 extern LTDC_HandleTypeDef hltdc;
